refactor: ladder.common 패키지명 attribute 로 수정
