+++ src/main/java/domain/ladder/LadderLeg.java
@@ -1,6 +1,6 @@
 package domain.ladder;
 
-import domain.ladder.common.Direction;
+import domain.ladder.attirbute.Direction;
 
 import java.util.List;
