+++ src/main/java/domain/ladder/LadderLegGenerator.java
@@ -1,7 +1,7 @@
 package domain.ladder;
 
-import domain.ladder.common.Direction;
-import domain.ladder.common.Height;
+import domain.ladder.attirbute.Direction;
+import domain.ladder.attirbute.Height;
 
 import java.util.List;
 import java.util.function.Supplier;
