+++ src/test/java/domain/ladder/LadderLegPieceTest.java
@@ -1,7 +1,6 @@
 package domain.ladder;
 
-import domain.ladder.common.Direction;
-import domain.ladder.LadderLegPiece;
+import domain.ladder.attirbute.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
